refactored the datatype of field displacement in all places

+++ src/main/java/com/engine/CNGEngine.java
@@ -13,7 +13,7 @@ public class CNGEngine extends Engine {
     public CNGEngine() {
     }
 
-    public CNGEngine(String modelNumber, float displacement, float maxHorsePower, int numberOfCylinders, int revolutionsPerMinute, String fuelType) {
+    public CNGEngine(String modelNumber, int displacement, float maxHorsePower, int numberOfCylinders, int revolutionsPerMinute, String fuelType) {
         super(modelNumber, displacement, maxHorsePower, numberOfCylinders, revolutionsPerMinute);
         this.fuelType = fuelType;
     }
