+++ src/main/java/com/engine/DieselEngine.java
@@ -13,7 +13,7 @@ public class DieselEngine extends Engine {
     public DieselEngine() {
     }
 
-    public DieselEngine(String modelNumber, float displacement, float maxHorsePower, int numberOfCylinders, int revolutionsPerMinute, String fuelType) {
+    public DieselEngine(String modelNumber, int displacement, float maxHorsePower, int numberOfCylinders, int revolutionsPerMinute, String fuelType) {
         super(modelNumber, displacement, maxHorsePower, numberOfCylinders, revolutionsPerMinute);
         this.fuelType = fuelType;
     }
