+++ src/main/java/com/engine/ElectricEngine.java
@@ -13,7 +13,7 @@ public class ElectricEngine extends Engine {
     public ElectricEngine() {
     }
 
-    public ElectricEngine(String modelNumber, float displacement, float maxHorsePower, int numberOfCylinders, int revolutionsPerMinute, float voltage) {
+    public ElectricEngine(String modelNumber, int displacement, float maxHorsePower, int numberOfCylinders, int revolutionsPerMinute, float voltage) {
         super(modelNumber, displacement, maxHorsePower, numberOfCylinders, revolutionsPerMinute);
         this.voltage = voltage;
     }
